# GetitDone/GetitDone.Service/generated/controllers/LabelOpsOperationsControllerBase.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Net;
using System.Threading.Tasks;
using System.Text.Json;
using System.Text.Json.Serialization;
using Microsoft.AspNetCore.Mvc;
using Getitdone.Service.Models;
using Getitdone.Service;

namespace Getitdone.Service.Controllers
{
    [ApiController]
    public abstract partial class LabelOpsOperationsControllerBase : ControllerBase
    {

        internal abstract ILabelOpsOperations LabelOpsOperationsImpl { get; }


        [HttpGet]
        [Route("/labels/{label_id}")]
        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(Label))]
        public virtual async Task<IActionResult> GetPersonalLabel(string labelId)
        {
            var result = await LabelOpsOperationsImpl.GetPersonalLabelAsync(labelId);
            return Ok(result);
        }


        [HttpPost]
        [Route("/labels/{label_id}")]
        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(Label))]
        public virtual async Task<IActionResult> UpdateLabel(string labelId, UpdateLabelRequest body)
        {
            var result = await LabelOpsOperationsImpl.UpdateLabelAsync(labelId, body);
            return Ok(result);
        }


        [HttpDelete]
        [Route("/labels/{label_id}")]
        [ProducesResponseType((int)HttpStatusCode.NoContent, Type = typeof(void))]
        public virtual async Task<IActionResult> DeleteLabel(string labelId)
        {
            await LabelOpsOperationsImpl.DeleteLabelAsync(labelId);
            return Ok();
        }

    }
}