# GetitDone/GetitDone.Service/generated/controllers/SharedLabelsOperationsControllerBase.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Net;
using System.Threading.Tasks;
using System.Text.Json;
using System.Text.Json.Serialization;
using Microsoft.AspNetCore.Mvc;
using Getitdone.Service.Models;
using Getitdone.Service;

namespace Getitdone.Service.Controllers
{
    [ApiController]
    public abstract partial class SharedLabelsOperationsControllerBase : ControllerBase
    {

        internal abstract ISharedLabelsOperations SharedLabelsOperationsImpl { get; }


        [HttpGet]
        [Route("/labels/shared")]
        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(Label[]))]
        public virtual async Task<IActionResult> GetSharedLabels()
        {
            var result = await SharedLabelsOperationsImpl.GetSharedLabelsAsync();
            return Ok(result);
        }


        [HttpPost]
        [Route("/labels/shared/rename")]
        [ProducesResponseType((int)HttpStatusCode.NoContent, Type = typeof(void))]
        public virtual async Task<IActionResult> RenameSharedLabel(RenameSharedLabelRequest body)
        {
            await SharedLabelsOperationsImpl.RenameSharedLabelAsync(body);
            return Ok();
        }


        [HttpPost]
        [Route("/labels/shared/remove")]
        [ProducesResponseType((int)HttpStatusCode.NoContent, Type = typeof(void))]
        public virtual async Task<IActionResult> RemoveSharedLabel(RemoveSharedLabelRequest body)
        {
            await SharedLabelsOperationsImpl.RemoveSharedLabelAsync(body);
            return Ok();
        }

    }
}