# GetitDone/servers/aspnet/generated/models/Collaborator.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace Getitdone.Service.Models
{

    public partial class Collaborator
    {
        public string Id { get; set; }

        public string Name { get; set; }

        public string Email { get; set; }


    }
}