# GetitDone/servers/aspnet/generated/models/TodoItem.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace Getitdone.Service.Models
{

    public partial class TodoItem
    {
        [JsonPropertyName("creator_id")]
        public string CreatorId { get; set; }

        [JsonPropertyName("created_at")]
        public string CreatedAt { get; set; }

        [JsonPropertyName("assignee_id")]
        public string AssigneeId { get; set; }

        [JsonPropertyName("assigner_id")]
        public string AssignerId { get; set; }

        [JsonPropertyName("comment_count")]
        public int CommentCount { get; set; }

        [JsonPropertyName("is_completed")]
        public bool IsCompleted { get; set; }

        public string Content { get; set; }

        public string Description { get; set; }

        public Due Due { get; set; }

        public string Duration { get; set; }

        public string Id { get; set; }

        public string[] Labels { get; set; }

        public int Order { get; set; }

        public int Priority { get; set; }

        [JsonPropertyName("project_id")]
        public string ProjectId { get; set; }

        [JsonPropertyName("section_id")]
        public string SectionId { get; set; }

        [JsonPropertyName("parent_id")]
        public string ParentId { get; set; }

        public string Url { get; set; }


    }
}